Extracted wagmi status context to new file
