+++ src/components/Layout/Navbar/MobileNav.tsx
@@ -22,8 +22,8 @@ import { mobileWalletDetails, MOBILE_NAV_ITEMS } from '@/data/NavItemData'
 import { MobileNavItem, MobileSubNav } from '@/components/Layout/Navbar'
 import NavSearch from '@/components/Layout/Navbar/NavSearch'
 import { getUserAddressFromCache } from '@/utils/getUserAddressFromCache'
-import { WagmiStatusContext } from '@/components/Wagmi/DynamicWagmiProvider'
 import dynamic from 'next/dynamic'
+import { WagmiStatusContext } from '@/components/Wagmi/WagmiStatusContext'
 import { ColorModeToggle } from './ColorModeToggle'
 
 const DeferredLogOutBtn = dynamic(() => import('./DeferredLogout'), {
